+++ packages/file-bundle/config/file-association.php
@@ -21,6 +21,7 @@ use Rekalogika\File\Association\ClassSignatureResolver\AttributeClassSignatureRe
 use Rekalogika\File\Association\ClassSignatureResolver\ChainClassSignatureResolver;
 use Rekalogika\File\Association\ClassSignatureResolver\DefaultClassSignatureResolver;
 use Rekalogika\File\Association\Contracts\ClassBasedFileLocationResolverInterface;
+use Rekalogika\File\Association\Contracts\ClassMetadataFactoryInterface;
 use Rekalogika\File\Association\Contracts\ClassSignatureResolverInterface;
 use Rekalogika\File\Association\Contracts\ObjectIdResolverInterface;
 use Rekalogika\File\Association\FilePropertyManager\DefaultFilePropertyManager;
@@ -72,7 +73,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->set('rekalogika.file.association.object_manager')
         ->class(DefaultObjectManager::class)
         ->args([
-            '$classMetadataFactory' => service('rekalogika.file.association.class_metadata_factory'),
+            '$classMetadataFactory' => service(ClassMetadataFactoryInterface::class),
             '$objectIdResolver' => service(ObjectIdResolverInterface::class),
             '$filePropertyManager' => service('rekalogika.file.association.file_property_manager'),
         ])
@@ -160,7 +161,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->set('rekalogika.file.association.class_based_file_location_resolver.default')
         ->class(DefaultClassBasedFileLocationResolver::class)
         ->args([
-            service('rekalogika.file.association.class_metadata_factory'),
+            service(ClassMetadataFactoryInterface::class),
         ])
         ->tag('rekalogika.file.association.class_based_file_location_resolver', [
             'priority' => -1000,
@@ -234,7 +235,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     //
 
     $services
-        ->set('rekalogika.file.association.class_metadata_factory')
+        ->set(ClassMetadataFactoryInterface::class)
         ->class(DefaultClassMetadataFactory::class)
         ->args([
             service('rekalogika.file.association.property_lister'),
@@ -251,7 +252,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services
         ->set('rekalogika.file.association.class_metadata_factory.caching')
         ->class(CachingClassMetadataFactory::class)
-        ->decorate('rekalogika.file.association.class_metadata_factory')
+        ->decorate(ClassMetadataFactoryInterface::class)
         ->args([
             service('.inner'),
             service('rekalogika.file.association.class_metadata_factory.cache'),